requirements.txt
python-chess>=1.999
numpy>=1.24.0
tensorflow>=2.13.0